requirements.txt
numpy
paddlepaddle_gpu>=2.2
Pillow
wandb
